backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
numpy==1.26.2
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
dj-database-url==2.1.0
